Fixed many bugs and added some features.
Fixed the bug where the Change Database button wouldn't reload the database properly if the app was already running.
Added a New Database Button.
Removed the search button and made the search box search as you type.
Made the search list a table instead of a wall of text that causes people's eyes to bleed.
Made the database view list a table instead of a wall of text that causes people's eyes to bleed.
Fixed all (I think) textboxes so that ctrl-backspace works and added watermark text to them.
Also made the text in all dropdown menus non-editable.