Fixed some bugs
Fixed a bug where pressing tab jumped around the page in a very random order for most things.
Fixed a bug where the add item page didn't completely clear after adding an item.
DBView and Search ListViews are now ascending/descending sortable by column. There is currently no indication of how it is sorted (like an arrow on the column), but it works.